fix date type error and make userprofile redirect
The date type will show different date from the data store in psql.
This is cause because of the node pg will get data in the timezone not
in Taipei.
So dierct use string slice will cause the problem.

make the redirect work and remove redundant call to check session data.